Experimenting with a bounded cache.
This commit uses an LRU bounded map from

https://github.com/blitzcode/lru-bounded-map

in order to test using Haxl with a bounded cache. While this removes the
function of the cache to ensure consistency, it allows to use Haxl with
a large number of requests that involve nontrivial amounts of data. With
the unbounded cache originally implemented in Haxl, we'd risk requiring
unbounded amounts of memory in that use case.